Fixes compile time error reported by anonymous on #3827. 64 bit safety for 32 bit library.
git-svn-id: http://svn.mythtv.org/svn/trunk@14292 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythplugins/mythmusic/mythmusic/aacdecoder.h
@@ -15,7 +15,6 @@
 
 #include <mp4ff.h>
 
-
 class Metadata;
 
 class aacDecoder : public Decoder
@@ -51,7 +50,7 @@ class aacDecoder : public Decoder
     bool done, finish;
     long len, bitrate;
     uchar channels;
-    unsigned long sample_rate;
+    uint32_t sample_rate;
     unsigned long output_size;
     double totalTime, seekTime;
